decode_exception: Don't use non-standard off_t type.
Also eliminates a truncation warning on MSVC.

+++ include/spotify/json/decode_exception.hpp
@@ -39,7 +39,7 @@ class decode_exception : public std::runtime_error {
   }
 
  private:
-  const off_t _offset;
+  const size_t _offset;
 };
 
 }  // namespace json
